#tf-data-service GetSnapshotSplit returns the split indexes.
The worker needs it to validate the split it receives and the split
it reads from files do not have gaps. For example, in the following
sequence:

   worker requests the kth split
-> dispatcher receives request
-> worker dies
-> worker restarts
-> worker reads from files and processes (k-1) splits
-> dispatcher writes the kth split
-> worker requests the next split
-> dispatcher returns the (k+1)th split

In the above sequence, the worker misses the kth split. If the worker
knows the split index, it will wait for the kth split to appear in the
file system when it receives the (k+1)th split.

PiperOrigin-RevId: 505144776