Avoid unnecessary polling in EventMgr.
The TF EventMgr lets you enqueue a std::function to be run when an se::Stream
finishes all the work that's currently enqueued on it.

It does this by creating se::Event's on the streams and periodically polling
all of them to see if they're completed.  This poll loop is very expensive for
some clients.

If you have two se::Event's enqueued on the same se::Stream and the first event
has not been hit yet, then you can be sure that the second one also hasn't been
hit: A Stream's work runs in strict FIFO order.

Previously EventMgr would check all of the events on every stream, doing
unnecessary work.  This CL changes it so it stops after the first event on a
stream that hasn't been hit yet.  If there are often multiple events pending on
a particular stream, this should save significant CPU.

While we're here, we also cleaned up EventMgr.  Previously it had additional
functionality about freeing tensors, but this was ripped out a while ago.
Cleaning this up allows us to simplify the code somewhat.

PiperOrigin-RevId: 506161538