Fix a bug in the validation_model build macro.
This build macro uses a temporary file named $(@D)/tmp, which is a problem, because this build macro is instantiated twice in the same package, and both of the generated rules are executed in parallel during the build, and both try to write to the same tmp file, and then both of them try to remove it.  This leads to one of the rules failing due to the file not being found, on account of it having been removed by the other rule.

The fix is to instead use $(@D)/<name>.tflite.tmp as the name of the temporary file, where <name> is the name of the rule.  This is sufficient to ensure that the temporary file names used by the different instantiations of this build macro are distinct.

PiperOrigin-RevId: 506882948