Add a new replica context for DTensor related strategy.
Since DTensor operate in a global context, we mostly raise an explicit error about methods are not available in the context. This will lead to a behavior discrepancy between the existing strategy and new one. We should consider this carefully for the future plan when rebasing the strategy on top of DTensor (eg using replicated_run/spmd_run).

PiperOrigin-RevId: 506945673