Speed up merge/purge in MemoryCheckpoint.
The idea is that we generate an int id for each string key in checkpoint and use ids as map keys in MemoryCheckpoint. The mapping between id and prefix is kept in a shared map created in IteratorResource. When purging a prefix, we remove all the key-value pairs where the key starts with the prefix (by scanning the key to id map).

PiperOrigin-RevId: 507590042