Remove tensor() from TensorWithLayout. It will only appear in TensorWithLayoutTf.
TensorWithLayout does not depend on parallel_device anymore.

The usages are all in the short-circuited code by ParallelExecutor. In case that one wants to get the underlying TensorHandle, get_tensor(index) is enough.

The returning TensorHandlePtr* implicitly assumes the handles are stored in a continuous memory space, which isn't specified in the contract.

PiperOrigin-RevId: 507611615