tf_gen_op_wrapper_py op_allowlist support for ops in namespaces
Ops in namespaces have ">" in their name. Quote names so ">" does not cause problems when calling the wrapper generator. (Note that this fix modifies one line in tensorflow.bzl, all rest of the additions to other files is for adding a test.)

Here's a BUILD file snippet illustrating what now works with this fix:

tf_gen_op_wrapper_py(
    # ...
    op_allowlist = [
        "ExampleNamespace>ExampleOp",
    ],
    # ...
)

PiperOrigin-RevId: 507627783