Add experimental `friendly_name` config option.
Today when you use XLA autoclustering, the XLA module names have the form e.g.

  cluster_15111732669523428041_0__XlaCompiledKernel_true__XlaHasReferenceVars_false__XlaNumConstantArgs_0__XlaNumResourceArgs_0_.2393

If you have a program with many TensorFlow nets, it is hard to tell which
cluster_1234 belongs to which net.

With this change, you can set a name in the TF session config and the name gets
propagated down to the XLA cluster name:

  my_friendly_name_15111732669523428041_0__<snip>

For now this only works for nets that enable XLA via autoclustering.  I'd like
to do something similar for nets that use tf.function and inference-converter.

PiperOrigin-RevId: 507687578