Implemented a pass that inserts a SaveV2Op in TF Quantizer.
This is a sibling pass with InsertRestoreOpPass. In order to correctly save the variable snapshots during model saving, there should be a SaveV2 op.
This new pass inserts a SaveV2 op for each VarHandleOp in the initializer function of type "restore_op".

InsertSaveOpPass is added as part of the constant unfreezing function but the in this change there should not be visible changes in the saving / restoring routine of the exported quantized model.

PiperOrigin-RevId: 507940347