Adjust attr summary for long lists to make it more human readable.
The previous attr summary included a fingerprint in the middle of the list
if the list size was too long.  This fingerprint is meant to ensure uniqueness, since
the string is used for hashing.  However, the op summary (including attrs) is often
printed to the user, including when input arguments are invalid, which can lead to
confusion.

Increased the max size of attr list to 30 so it can display all current allowed dtype
lists without being compressed, replaced fingerprint with ellipses to make it more
human-readable, and added a separate `{attr_hash=...}` string with the fingerprint
to restore uniqueness.

Fixes #58696.

PiperOrigin-RevId: 508091035