Warn and continue on GIF errors if there is at least one image read.
Some GIF images seem to contain extraneous data after the image section that causes
the giflib library to fail to fully parse the file.  It is unclear whether
this is actually a bug in giflib, or if other image libraries like PIL
simply ignore the offending data.  Here, if at least one image is successfully read but
the full GIF file fails to parse, we log the error but continue loading the image.

PiperOrigin-RevId: 508144639