Migrate XlaCallModuleOp to --stablehlo-shape-refinement
Specialization of dynamically-shaped programs to static shapes is one of the key requirements of XlaCallModuleOp, and in the last two quarters we've got a lot of mileage out of --tf-shape-inference. Even though this pass was designed for a different use case, it fit our use case surprisingly well and unblocked experiments without requiring much custom development.

Earlier this year, we have implemented --stablehlo-refine-shapes, a pass that propagates shape information throughout StableHLO programs. This pass is built to accommodate all the peculiarities of StableHLO type inference, and having gone through the entirety of the StableHLO opset, I am confident that it represents a rock-solid foundation for XlaCallModuleOp.

Future work:
  * Design a mechanism to customize shape inference for CustomCallOp: https://github.com/openxla/stablehlo/issues/851.
  * Implement shape refinement for DynamicConvOp (https://github.com/openxla/stablehlo/issues/861).
  * Implement type inference for AllGatherOp (https://github.com/openxla/stablehlo/issues/865) and ReduceScatterOp (https://github.com/openxla/stablehlo/issues/864).
  * Generalize the pass to be applicable to MHLO (https://github.com/openxla/stablehlo/issues/1160).

PiperOrigin-RevId: 509320369