Use disable_optimize_for_static_graph() in Tensorflow's GraphExecutionState.
This change ignores the value of the flag optimize_for_static_graph() and
replaces it by disable_optimize_for_static_graph().

As a result, this will enable optimizations for the static graph by default.
The optimization saves memory by discarding GraphDef protobufs that are given
to the Session upon creation.

Note that this optimization is incompatible with the API Session::Extend().
Therefore, any calls to the Session::Extend() API will need to be preceded by
a call to set the flag disable_optimize_for_static_graph() to true.

PiperOrigin-RevId: 509369591