Optimize 210 transposes.
Similarly as for 021 transposes which swap the last two dimensions, we can use the
tiled transpose emitter for 210 transposes which swap the first and the last dimension.
This requires a few adjustments, because so far it was assumed that the shared
memory tiles always consist of the last two dimensions, but for 210 transpose,
we want them to consist of the first and last dimension.

PiperOrigin-RevId: 510976080