Add upper bound to `wrapt`.
This because we detected that `wrapt==1.15.0rc0` is incompatible with `tensorflow==2.12.0rc0`. In fact, the incompatibility is present in nightly too, but because there we only test with released deps, the issue was not discovered. For RC0, a test script had `pip install --pre` which would install pre-release versions __for all packages__ that are present (security issue in general, but helped us here).

In order to not have TF 2.12 broken weeks after release when wrapt finalizes the 1.15 release, adding an upper bound now as preemptive.

This is first in master branch and will need to be cherrypicked.

This is a temporary upper bound, we need to understand why the compatibility issue arises and solve it, either in TF side or on wrapt side, or both.

PiperOrigin-RevId: 511219724