dtensor.run_on affects tf.function.
The mesh is used as the default mesh when no other meshes can be determined.
This improves the consistency and actually fixes a bug in DTensorMirroredStrategy's
run() function when the payload func is nullary.

Also fixed fragile mesh_propagation.mlir test, which was using CHECK-LABEL
with non-unique names, causing the split tests to be order dependent.

PiperOrigin-RevId: 512682063