Relocate default conversion registration to `constant_op.py`.
Adds a new internal-only method to `tensor_conversion_registry.py` which
allows the registration of tensor conversions even for the so-called
"unconvertible" types which cannot have conversions registered using the
exported method.

This method is named `register_tensor_conversion_function_internal` and is
unexported. It has the same arguments as `register_tensor_conversion_function`
and is otherwise identical in functionality to the exported version.

Removes the code from `tensor_conversion_registry` which unilaterally deferred to
`constant_op.constant()` as the conversion function for the unconvertible types.

Adds code to `constant_op.py` to register `constant_op.constant()` as the
conversion function for these types instead.

This has the side effect that conversions of these types will fail
unless the `constant_op.py` module has been loaded - this was true before,
but the (removed) lazy-load in `tensor_conversion_registry.py` enforced this.

PiperOrigin-RevId: 512957513