assertProtoEqual calls checkFloatEqAndReplace which recurses the proto and compares all floats/doubles using relative tolerance.
The current version's comparison is not valid for all floats, for example we probably want 50000 and 49999.9 to compare equal but they won't in the original version as the float format strings cut digits, it doesn't round so could result in 50000 and 49999, depending on the formatting provided.

This means that tests will be less brittle as a single different bit will not cause a failure if rtol is used.

PiperOrigin-RevId: 512999626