Handle `BatchFunctionOp` when analyzing whether a resource is potentially written.
`BatchFunctionOp` is not currently handled in the resource analyzer. Even though it looks like a function call, it is not actually implementing `CallOpInterface`, so any `VarHandleOp`s that are used by `BatchFunctionOp`s are considered mutated.
This change adds `BatchFunctionOp` as a special case for the resource analyzer so that if the callee does not mutate the variables then it is not considered "potentially written".

PiperOrigin-RevId: 513667001