Take FLR and fhandle from the last task in `BatchFunctionKernel`
This CL changes `BatchFunctionKernel` to take the `FunctionLibraryRuntime` and the function handle for batched execution from the last batch task, instead of those captured at the `BatchResource` creation time. This is to address an issue where FLR provided at the time of `BatchResource` creation may not be alive throughout the `BatchResource`'s lifetime. For example, eager operations use overlay FLR, which is refcounted.

Since the above requires providing the fhandle while registering inputs, `BatchResourceBase::RegisterInput` has been extended to take a function that constructs a batch task to be used. This is to avoid encoding fhandle as `RegisterInput`'s argument, which is specific to `BatchFunction`'s implementation, e.g., it does not apply to the TFRT fallback version of `BatchFunctionFallbackKernel`.

The implementation in this CL is strictly safer than the current one because FLR and fhandle are always captured together from the same context (for the last batch task), which is always guaranteed to be alive. For example, taking only FLR from the last task and providing fhandle at resource construction time might have simplified the implementation, but it's potentially less safe because such an approach has a risk of fhandle vs. FLR mismatch.

PiperOrigin-RevId: 513893191