Add BufferedWritableFile for buffered writing.
We have a few places where buffered writing is useful, however we do not have a
general buffered file class. This moves the logic from tensor_bundle and makes
it reusable elsewhere in TF.  The default file interface is unbuffered and for
some use cases (remote files + small writes), this results in major slowdowns.

PiperOrigin-RevId: 513969757