Simplify the XlaCallModuleOp to always do shape refinement, not only when the arguments have dynamic shapes.
This has several advantages. First, it simplifies the code, there are fewer distinct code paths to keep track of. Second, this is handling the case when there are dynamic shapes in the body of the function even if the arguments have static shapes.

I have also improved an error message when the module is invoked with the wrong number of arguments.

PiperOrigin-RevId: 514415781