Enable XlaCallModule to work with modules with dynamic shapes but no dim_args_spec
After the change in cl/514075116 when we run shape refinement in all cases, all that was needed was to move the inlining from the end of the AddMainWrapper (which is only called if we have dim_args_spec) to the start of RefineShapes. We need to call the inliner to remove calls to functions that may be called with the input arguments. Otherwise, when we refine the shapes of the input arguments to have static shapes we may create an invalid MLIR module.

PiperOrigin-RevId: 514422322